dev_requirements.txt
black==19.3b0
mypy==0.720
pytest==5.0.1
pytest-mock==1.10.4
pytest-postgresql==2.1.0
sqlalchemy-stubs
